# src/main/resources/static/swagger-ui/openapi3.yaml
openapi: 3.0.1
info:
  title: 801base API docs
  description: 801base의 API 문서입니다.
  version: 0.0.1
servers:
- url: https://api2.801base.com
...
paths:
...
  /reviews:
    post:
      tags:
      - Review
      summary: 배송완료된 상품에 리뷰를 등록할 수 있다.
      description: |2-
                        RVW-001: 리뷰 내용을 입력해주세요.
                        RVW-002: 별점은 1~5점 사이로 선택해주세요.
                        RVW-003: 배송완료된 주문건에 대해서만 리뷰를 작성할 수 있습니다.
                        RVW-004: 배송완료 후 30일 이내의 주문건에 대해서만 리뷰를 작성 및 수정할 수 있습니다.
                        RVW-005: 리뷰는 주문 당 한번만 작성가능합니다.
      operationId: 리뷰_등록_성공
      parameters:
      - name: Authorization
        in: header
        description: Authorization
        required: true
        schema:
          type: string
        example: Bearer sample-token
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/reviews1890941445"
            examples:
              리뷰_등록_성공:
                value: |-
                  {
                    "orderNumber" : "ORDER1234",
                    "orderItemId" : 10,
                    "rating" : 3,
                    "content" : "좋습니다."
                  }
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/reviews-1411419170"
              examples:
                리뷰_등록_성공:
                  value: |-
                    {
                      "data" : {
                        "reviewId" : 10
                      },
                      "error" : null
                    }
...
  /reviews/{reviewId}:
    put:
      tags:
      - Review
      summary: 리뷰를 수정할 수 있다.
      description: |2-
                        RVW-001: 리뷰 내용을 입력해주세요.
                        RVW-002: 별점은 1~5점 사이로 선택해주세요.
                        RVW-006: 리뷰를 찾을 수 없습니다.
                        RVW-007: 다른 사람의 리뷰를 수정할 수 없습니다
      operationId: 리뷰_수정_성공
      parameters:
      - name: reviewId
        in: path
        description: 리뷰 아이디
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        description: Authorization
        required: true
        schema:
          type: string
        example: Bearer sample-token
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/reviews-reviewId480882763"
            examples:
              리뷰_수정_성공:
                value: |-
                  {
                    "rating" : 3,
                    "content" : "좋습니다."
                  }
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/reviews-reviewId-215377401"
              examples:
                리뷰_수정_성공:
                  value: |-
                    {
                      "data" : {
                        "reviewId" : 10
                      },
                      "error" : null
                    }
    delete:
      tags:
      - Review
      summary: 리뷰를 삭제할 수 있다.
      description: |2-
                        RVW-006: 리뷰를 찾을 수 없습니다.
                        RVW-008: 다른 사람의 리뷰를 삭제할 수 없습니다
      operationId: 리뷰_삭제_성공
      parameters:
      - name: reviewId
        in: path
        description: 리뷰 아이디
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        description: Authorization
        required: true
        schema:
          type: string
        example: Bearer sample-token
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/admin-reviews-reviewId-reply1686665306"
              examples:
                리뷰_삭제_성공:
                  value: |-
                    {
                      "data" : {
                        "message" : "OK"
                      },
                      "error" : null
                    }
...
  /user/me:
    get:
      tags:
      - User
      summary: 내 정보조회
      operationId: 내_정보조회_성공
      parameters:
      - name: Authorization
        in: header
        description: Authorization
        required: true
        schema:
          type: string
        example: Bearer sample-token
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user-me-1207955323"
              examples:
                내_정보조회_성공:
                  value: |-
                    {
                      "data" : {
                        "name" : "홍길동",
                        "email" : "길동@닷.컴",
                        "nickname" : "동길홍ㅋㅋ"
                      },
                      "error" : null
                    }
    patch:
      tags:
      - User
      summary: 내 정보수정
      operationId: 내_정보수정_성공
      parameters:
      - name: Authorization
        in: header
        description: Authorization
        required: true
        schema:
          type: string
        example: Bearer sample-token
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/user-me1642029314"
            examples:
              내_정보수정_성공:
                value: |-
                  {
                    "nickname" : "하하하"
                  }
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/admin-orders-orderId-cancel486549215"
              examples:
                내_정보수정_성공:
                  value: |-
                    {
                      "data" : null,
                      "error" : null
                    }
...
components:
  schemas:
    reviews-1411419170:
      type: object
      properties:
        data:
          required:
          - reviewId
          type: object
          properties:
            reviewId:
              type: number
              description: 생성된 리뷰 아이디
    user-me-1207955323:
      type: object
      properties:
        data:
          required:
          - email
          - name
          - nickname
          type: object
          properties:
            nickname:
              type: string
              description: 닉네임
            name:
              type: string
              description: 이름
            email:
              type: string
              description: 이메일
...
    reviews-reviewId-215377401:
      type: object
      properties:
        data:
          required:
          - reviewId
          type: object
          properties:
            reviewId:
              type: number
              description: 수정된 리뷰 아이디
...
    reviews1890941445:
      required:
      - content
      - orderItemId
      - orderNumber
      - rating
      type: object
      properties:
        orderNumber:
          type: string
          description: 주문번호
        orderItemId:
          type: number
          description: 주문 아이템 아이디
        rating:
          type: number
          description: 별점
        content:
          type: string
          description: 리뷰 내용
...
    admin-reviews-reviewId-reply1686665306:
      type: object
      properties:
        data:
          required:
          - message
          type: object
          properties:
            message:
              type: string
              description: 응답 메시지
    user-me1642029314:
      required:
      - nickname
      type: object
      properties:
        nickname:
          type: string
          description: 닉네임
...
    admin-orders-orderId-cancel486549215:
      type: object
...
    reviews-reviewId480882763:
      required:
      - content
      - rating
      type: object
      properties:
        rating:
          type: number
          description: 별점
        content:
          type: string
          description: 리뷰 내용
...